refactor: Answers 를 별도의 객체로 분리
