+++ src/main/java/baseball/Application.java
@@ -3,7 +3,6 @@ package baseball;
 public class Application {
 	public static void main(String[] args) {
 		NumberBaseballGame numberBaseballGame = new NumberBaseballGame();
-		numberBaseballGame.initAnswers();
 		numberBaseballGame.start();
 	}
 }
